+++ src/aux-server/CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### Changes:
 - Improvements
-  - Added top-level loading screen to Aux Builder.
+  - Added top-level loading screen to Aux Builder and Aux Player.
   
 ## V0.4.1
 ### Date: 4/05/2019